requirements.txt
dnspython==2.6.1
python-whois==0.8.0
aiohttp==3.8.5
beautifulsoup4==4.12.2
cryptography==41.0.4
aiodns==3.1.1
pycares==4.3.0
lxml==4.9.3